reorganize appswitcher, make it actually switch apps on click
